Make esbuild-file-plugin work in both dev/prod and rename it to esbuild-file-url-plugin
