+++ tools/esbuild-file-url-plugin/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "esbuild-file-plugin",
+  "name": "esbuild-file-url-plugin",
   "version": "0.0.0",
   "main": "src/index.js",
   "devDependencies": {
