+++ tools/yas-build-tools/package.json
@@ -9,7 +9,7 @@
     "@vanilla-extract/webpack-plugin": "^2.3.0",
     "@vitejs/plugin-react": "4.0.1",
     "@yas/env": "workspace:*",
-    "esbuild-file-plugin": "workspace:*",
+    "esbuild-file-url-plugin": "workspace:*",
     "css-minimizer-webpack-plugin": "^5.0.1",
     "rollup-plugin-visualizer": "5.9.0",
     "tsup": "7.2.0",
